requirements.txt
mysqlclient==2.0.*
python-telegram-bot==13.3
pytz
requests<=3.0.0
sqlalchemy==1.4.0b3
yoyo-migrations==7.3.*
